Changed the file structure so that the code actually works by itself.
It seems that you are using the /src folder to store the mod. However,
that makes it hard to encapsulate mods. Instead, I recommend you do the
following:
1. Back up all the files you are working on that are different from
their Github versions.
2. Delete your local Civilization Now repo from your computer.
3. Clone the online Civilization Now repo into the folder where your dev
enviroment (not /eclipse, but its parent).
4. Open eclipse and set the repo folder as a source folder.

Now you have all the content of this mod in its own separate folder.